search document 수정 (#58)
